# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hotels/{id}": {
      "get": {
        "operationId": "HotelController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel retrieved successfully"
          },
          "404": {
            "description": "Hotel not found"
          }
        },
        "summary": "Get hotel by ID",
        "tags": [
          "Hotels"
        ]
      },
      "patch": {
        "operationId": "HotelController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHotelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hotel updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Hotel not found"
          }
        },
        "summary": "Update hotel (Admin only)",
        "tags": [
          "Hotels"
        ]
      },
      "delete": {
        "operationId": "HotelController_deleteHotel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Hotel not found"
          }
        },
        "summary": "Delete hotel (Admin only)",
        "tags": [
          "Hotels"
        ]
      }
    },
    "/hotel-bookings": {
      "post": {
        "operationId": "HotelBookingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hotel booking created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Hotel not found"
          }
        },
        "summary": "Create a new hotel booking",
        "tags": [
          "Hotel Bookings"
        ]
      },
      "get": {
        "operationId": "HotelBookingController_findAll",
        "parameters": [
          {
            "name": "customer_name",
            "required": false,
            "in": "query",
            "description": "Filter by customer name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_email",
            "required": false,
            "in": "query",
            "description": "Filter by customer email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotel_id",
            "required": false,
            "in": "query",
            "description": "Filter by hotel ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_guests",
            "required": false,
            "in": "query",
            "description": "Filter by minimum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_guests",
            "required": false,
            "in": "query",
            "description": "Filter by maximum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by minimum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by maximum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "description": "Sort order (asc/desc)",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel bookings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get all hotel bookings with filters",
        "tags": [
          "Hotel Bookings"
        ]
      }
    },
    "/hotel-bookings/my-bookings": {
      "get": {
        "operationId": "HotelBookingController_findMyBookings",
        "parameters": [
          {
            "name": "customer_name",
            "required": false,
            "in": "query",
            "description": "Filter by customer name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_email",
            "required": false,
            "in": "query",
            "description": "Filter by customer email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotel_id",
            "required": false,
            "in": "query",
            "description": "Filter by hotel ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_guests",
            "required": false,
            "in": "query",
            "description": "Filter by minimum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_guests",
            "required": false,
            "in": "query",
            "description": "Filter by maximum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by minimum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by maximum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "description": "Sort order (asc/desc)",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User hotel bookings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get current user hotel bookings",
        "tags": [
          "Hotel Bookings"
        ]
      }
    },
    "/hotel-bookings/pending": {
      "get": {
        "operationId": "HotelBookingController_findPendingBookings",
        "parameters": [
          {
            "name": "customer_name",
            "required": false,
            "in": "query",
            "description": "Filter by customer name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_email",
            "required": false,
            "in": "query",
            "description": "Filter by customer email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotel_id",
            "required": false,
            "in": "query",
            "description": "Filter by hotel ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_guests",
            "required": false,
            "in": "query",
            "description": "Filter by minimum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_guests",
            "required": false,
            "in": "query",
            "description": "Filter by maximum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by minimum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by maximum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "description": "Sort order (asc/desc)",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending hotel bookings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get all pending hotel bookings",
        "tags": [
          "Hotel Bookings"
        ]
      }
    },
    "/hotel-bookings/hotel/{hotelId}": {
      "get": {
        "operationId": "HotelBookingController_findByHotel",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_name",
            "required": false,
            "in": "query",
            "description": "Filter by customer name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_email",
            "required": false,
            "in": "query",
            "description": "Filter by customer email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotel_id",
            "required": false,
            "in": "query",
            "description": "Filter by hotel ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-in date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_from",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_out_date_to",
            "required": false,
            "in": "query",
            "description": "Filter by check-out date to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_guests",
            "required": false,
            "in": "query",
            "description": "Filter by minimum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_guests",
            "required": false,
            "in": "query",
            "description": "Filter by maximum guests",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by minimum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_rooms",
            "required": false,
            "in": "query",
            "description": "Filter by maximum rooms",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "description": "Sort order (asc/desc)",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel bookings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get hotel bookings for a specific hotel",
        "tags": [
          "Hotel Bookings"
        ]
      }
    },
...
    "/hotel-bookings/{id}/approve": {
      "patch": {
        "operationId": "HotelBookingController_approveBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel booking approved successfully"
          },
          "400": {
            "description": "Bad request - booking cannot be approved"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pending booking not found"
          }
        },
        "summary": "Approve a pending hotel booking",
        "tags": [
          "Hotel Bookings"
        ]
      }
    },
    "/hotel-bookings/{id}/reject": {
      "patch": {
        "operationId": "HotelBookingController_rejectBooking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel booking rejected successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pending booking not found"
          }
        },
        "summary": "Reject a pending hotel booking",
        "tags": [
          "Hotel Bookings"
        ]
      }
    }
  },
  "info": {
    "title": "Travel Booking API Documentation",
    "description": "API documentation for Travel Booking",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "[just text field] Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
...
  },
  "security": [
    {
      "jwt": []
    }
  ]
}